# Conflicts:
#	.idea/modules/core/core.iml
#	.idea/modules/core/core_main.iml
#	.idea/modules/core/core_test.iml
#	.idea/modules/desktop/desktop.iml
#	.idea/modules/desktop/desktop_main.iml
#	.idea/modules/desktop/desktop_test.iml
#	assets/session.dat

+++ core/src/com/rastera/hubg/Sprites/InteractiveTileObject.java
@@ -23,11 +23,11 @@ public abstract class InteractiveTileObject {
         PolygonShape shape = new PolygonShape();
 
         bdef.type= BodyDef.BodyType.StaticBody;
-        bdef.position.set((float) (bounds.getX() + bounds.getWidth() / 2), (float) (bounds.getY() + bounds.getHeight() / 2));
+        bdef.position.set(bounds.getX() + bounds.getWidth() / 2, bounds.getY() + bounds.getHeight() / 2);
 
         body = world.createBody(bdef);
 
-        shape.setAsBox((float)bounds.getWidth(), (float) bounds.getHeight());
+        shape.setAsBox(bounds.getWidth(), bounds.getHeight());
         fdef.shape = shape;
 
         body.createFixture(fdef);
